feature
finished person activity

+++ TestCourseWork/app/src/main/java/com/example/testcoursework/model/data/dataClass/Person.kt
@@ -2,4 +2,4 @@ package com.example.testcoursework.model.data.dataClass
 
 import androidx.lifecycle.MutableLiveData
 
-data class Person(var gender: MutableLiveData<String>, var height: MutableLiveData<Int>, var weight: MutableLiveData<Float>)
+data class Person(var gender: MutableLiveData<String>, var height: MutableLiveData<Int>, var weight: MutableLiveData<Int>)
